+++ scripts/prepare_sector_network.py
@@ -1483,6 +1483,108 @@ def cycling_shift(df, steps=1):
     return df
 
 
+def _add_other_non_res_tyndp(
+    n: pypsa.Network,
+    generator: str,
+    carrier: str,
+    carrier_nodes: list,
+    spatial: SimpleNamespace,
+    costs: pd.DataFrame,
+    pemmdb_capacities: pd.DataFrame,
+    co2_price: float,
+) -> None:
+    """
+    Add TYNDP Other Non-RES price bands to the network.
+
+    Creates links between carrier buses and demand nodes for Other Non-RES generators of different price bands,
+    including their efficiency, costs, and CO2 emissions.
+
+    Parameters
+    ----------
+    n : pypsa.Network
+        The PyPSA network container object
+    generator : str
+        Name of the Other Non-RES generator
+    carrier : str
+        Name of the Other Non-RES fuel carrier
+    carrier_nodes : list
+        Nodes of the fuel carrier
+    spatial : SimpleNamespace
+        Namespace containing spatial information for different carriers,
+        including nodes and locations
+    costs : pd.DataFrame
+        DataFrame containing cost and technical parameters for different technologies
+    pemmdb_capacities : pd.DataFrame
+        Dataframe containing PEMMDB capacities including information on the different Other Non-RES price bands
+    co2_price: float
+        Emission price for the given planning year
+
+    Returns
+    -------
+    None
+        Modifies the network object in-place by adding TYNDP Other Non-RES price bands
+    """
+    query = "" if "ccs" in generator else "not"
+
+    price_bands = pemmdb_capacities.query(
+        f"index_carrier.str.startswith(@generator) and {query} index_carrier.str.contains('ccs') and efficiency > 0"
+    ).reset_index()
+    nodes = price_bands.bus.values
+    offer_price = price_bands.price
+    fuel_price = costs.at[carrier, "fuel"]
+    adjusted_price = (
+        offer_price * price_bands.efficiency  # EUR/MWhel * efficiency = EUR/MWhth
+        - fuel_price  # EUR/MWhth
+        - price_bands.co2_factor * co2_price  # tCO2/MWhth * EUR/tCO2 = EUR/MWhth
+    ).values
+
+    if "ccs" in generator:
+        capture_rate = costs.at[generator, "capture_rate"]
+        co2_capture = capture_rate * price_bands.co2_factor.values / (1 - capture_rate)
+
+        n.add(
+            "Link",
+            price_bands["bus"].astype(str)
+            + "-"
+            + price_bands["index_carrier"].astype(str)
+            + "-"
+            + price_bands["price"].round(2).astype(str)
+            + "eur",
+            bus0=carrier_nodes,
+            bus1=nodes,
+            bus2="co2 atmosphere",
+            bus3=spatial.co2.df.loc[nodes, "nodes"].values,
+            carrier=generator,
+            p_nom_extendable=False,
+            marginal_cost=adjusted_price,
+            efficiency=price_bands.efficiency.values,
+            efficiency2=price_bands.co2_factor.values,
+            efficiency3=co2_capture,
+            lifetime=costs.at[generator, "lifetime"],
+        )
+
+    # Else add as regular conventional power plants
+    else:
+        n.add(
+            "Link",
+            price_bands["bus"].astype(str)
+            + "-"
+            + price_bands["index_carrier"].astype(str)
+            + "-"
+            + price_bands["price"].round(2).astype(str)
+            + "eur",
+            bus0=carrier_nodes,
+            bus1=nodes,
+            bus2="co2 atmosphere",
+            carrier=generator,
+            p_nom_extendable=False,
+            marginal_cost=adjusted_price,
+            efficiency=price_bands.efficiency.values,
+            efficiency2=price_bands.co2_factor.values,
+            lifetime=costs.at[generator, "lifetime"],
+        )
+
+
 def add_thermal_generation_tyndp(
     n: pypsa.Network,
     costs: pd.DataFrame,
@@ -1491,6 +1593,8 @@ def add_thermal_generation_tyndp(
     spatial: SimpleNamespace,
     options: dict,
     cf_industry: dict,
+    pemmdb_capacities: pd.DataFrame,
+    co2_price: float,
 ) -> None:
     """
     Add TYNDP conventional thermal electricity generation technologies to the network.
@@ -1515,6 +1619,10 @@ def add_thermal_generation_tyndp(
         Configuration dictionary containing settings for the model
     cf_industry : dict
         Dictionary of industrial conversion factors, needed for carrier buses
+    pemmdb_capacities: pd.DataFrame
+        Dataframe containing PEMMDB capacities including Other Non-RES price band information
+    co2_price: float
+        Emission price for the given planning year
 
     Returns
     -------
@@ -1540,9 +1648,22 @@ def add_thermal_generation_tyndp(
             cf_industry=cf_industry,
         )
 
-        # Add CCS conventional power plants using default PyPSA-Eur assumptions for capture rates and capital cost
+        # Add TYNDP Other Non-RES price bands to the network
+        if "other-non-res" in generator:
+            _add_other_non_res_tyndp(
+                n=n,
+                generator=generator,
+                carrier=carrier,
+                carrier_nodes=carrier_nodes,
+                spatial=spatial,
+                costs=costs,
+                pemmdb_capacities=pemmdb_capacities,
+                co2_price=co2_price,
+            )
+
+        # Otherwise, add CCS conventional power plants using default PyPSA-Eur assumptions for capture rates and capital cost
         # TODO: Update with TYNDP specific assumptions
-        if "ccs" in generator:
+        elif "ccs" in generator:
             n.add(
                 "Link",
                 nodes + " " + generator,
@@ -1583,6 +1704,7 @@ def add_other_res_tyndp(
     costs: pd.DataFrame,
     pop_layout: pd.DataFrame,
     spatial: SimpleNamespace,
+    options: dict,
 ) -> None:
     """
     Add Other RES technologies to the network. This includes two groups of plants:
@@ -1601,6 +1723,8 @@ def add_other_res_tyndp(
     spatial : SimpleNamespace
         Namespace containing spatial information for different carriers,
         including nodes and locations.
+    options : dict
+        Configuration dictionary containing settings for the model.
 
     Returns
     -------
@@ -1611,16 +1735,27 @@ def add_other_res_tyndp(
 
     nodes = pop_layout.index
 
-    # Add Other RES Biomass as Links
-    n.add(
-        "Link",
-        nodes + " other-res-biomass",
-        bus0=spatial.biomass.df.loc[nodes, "nodes"].values,
-        bus1=nodes,
-        carrier="other-res-biomass",
-        p_nom_extendable=False,
-        efficiency=costs.at["central solid biomass CHP", "efficiency"],
-    )
+    # Add biomass components to the model incl. biomass bus
+    # TODO: refactor with add_carrier_buses both here and in add_biomass
+    if options["biomass"]:
+        n.add(
+            "Bus",
+            spatial.biomass.nodes,
+            location=spatial.biomass.locations,
+            carrier="solid biomass",
+            unit="MWh_LHV",
+        )
+
+        # Add Other RES Biomass as Links
+        n.add(
+            "Link",
+            nodes + " other-res-biomass",
+            bus0=spatial.biomass.df.loc[nodes, "nodes"].values,
+            bus1=nodes,
+            carrier="other-res-biomass",
+            p_nom_extendable=False,
+            efficiency=costs.at["central solid biomass CHP", "efficiency"],
+        )
 
     # Add Other RES Mix as Generators
     n.add(
@@ -1644,6 +1779,8 @@ def add_generation(
     ext_carriers: dict[str, list[str]],
     existing_capacities: pd.Series | None = None,
     existing_efficiencies: pd.Series | None = None,
+    pemmdb_capacities: pd.DataFrame | None = None,
+    co2_price: float | None = None,
 ) -> None:
     """
     Add conventional electricity generation to the network.
@@ -1677,6 +1814,10 @@ def add_generation(
         Existing generation capacities by generator type
     existing_efficiencies : pd.Series | None
         Existing generation efficiencies by generator type
+    pemmdb_capacities: pd.DataFrame | None
+        Dataframe containing PEMMDB capacities including Other Non-RES price band information
+    co2_price: float | None
+        Emission price for the given planning year
 
     Returns
     -------
@@ -1774,7 +1915,101 @@ def add_generation(
         spatial=spatial,
         options=options,
         cf_industry=cf_industry,
+        pemmdb_capacities=pemmdb_capacities,
+        co2_price=co2_price,
     )
+
+
+def _add_pu_profiles(
+    comp_t: dict[str, pd.DataFrame], p_min_pu: pd.DataFrame, p_max_pu: pd.DataFrame
+) -> None:
+    """
+    Safely add new per unit profiles while safeguarding against profiles with only default values.
+    """
+    # Drop columns that are all zeros (default-value)
+    p_min_pu = p_min_pu.loc[:, (p_min_pu != 0.0).any()]
+    # Drop columns that are all ones (default-value)
+    p_max_pu = p_max_pu.loc[:, (p_max_pu != 1.0).any()]
+
+    _add_new_profiles_to_existing(
+        component_t=comp_t, attr="p_min_pu", new_profiles=p_min_pu
+    )
+    _add_new_profiles_to_existing(
+        component_t=comp_t, attr="p_max_pu", new_profiles=p_max_pu
+    )
+
+
+def _add_other_non_res_capacities(
+    n: pypsa.Network,
+    pemmdb_capacities: pd.DataFrame,
+    pemmdb_profiles: pd.DataFrame,
+    tech: str,
+) -> None:
+    """
+    Add PEMMDB capacities and profiles to Other Non-RES price band components.
+
+    Parameters
+    ----------
+    n : pypsa.Network
+        The PyPSA network container object.
+    pemmdb_capacities : pd.DataFrame
+        All PEMMDB capacities.
+    pemmdb_profiles : pd.DataFrame
+        All PEMMDB must-run and availability profiles.
+    tech : str
+        Other Non-RES price band to be added to the network.
+
+    Returns
+    -------
+    None
+        Modifies the network object in-place by adding Other Non-RES capacities and profiles.
+    """
+
+    tech_i = n.links.query("carrier == @tech").index
+    if tech_i.empty:
+        return
+
+    # Capacities
+    ############
+    # Filter for capacities and add to the network
+    caps = (
+        pemmdb_capacities.query("open_tyndp_type == @tech")
+        .reset_index()
+        .assign(
+            price_band=lambda df: df["bus"].astype(str)
+            + "-"
+            + df["index_carrier"].astype(str)
+            + "-"
+            + df["price"].round(2).astype(str)
+            + "eur"
+        )
+        .set_index("price_band")
+    )["p_nom"]
+    # Existing capacities are given in MWel, hence we need to convert to MWth
+    n.links.loc[tech_i, "p_nom"] = (
+        caps.reindex(tech_i).fillna(0.0).div(n.links.loc[tech_i, "efficiency"])
+    )
+
+    # Profiles
+    ##########
+    # Filter for profiles
+    profiles = pemmdb_profiles.query("open_tyndp_type == @tech").assign(
+        link_index=lambda df: df["bus"].astype(str)
+        + "-"
+        + df["index_carrier"].astype(str)
+        + "-"
+        + df["price"].round(2).astype(str)
+        + "eur"
+    )
+    p_min_pu = profiles.pivot_table(
+        values="p_min_pu", index="time", columns="link_index"
+    ).reindex(tech_i, axis=1, fill_value=0.0)
+
+    p_max_pu = profiles.pivot_table(
+        values="p_max_pu", index="time", columns="link_index"
+    ).reindex(tech_i, axis=1, fill_value=1.0)
+
+    _add_pu_profiles(comp_t=n.links_t, p_min_pu=p_min_pu, p_max_pu=p_max_pu)
 
 
 def _add_conventional_thermal_capacities(
@@ -1802,12 +2037,21 @@ def _add_conventional_thermal_capacities(
         Trajectories for exogenous nuclear pathways.
     nuclear_profiles : pd.DataFrame
         DataFrame containing the availability profiles of nuclear power plants.
+
+    Returns
+    -------
+    None
+        Modifies the network object in-place by adding conventional thermal capacities and profiles.
     """
     logger.info(
         "Adding PEMMDB capacities and profiles to conventional thermal generation assets."
     )
 
     for tech in tyndp_conventional_thermals:
+        if "other-non-res" in tech:
+            _add_other_non_res_capacities(n, pemmdb_capacities, pemmdb_profiles, tech)
+            continue
+
         # Capacities
         ############
         # Get indices for this technology
@@ -1884,8 +2128,6 @@ def _add_conventional_thermal_capacities(
                 p_min_pu[common_cols] = np.minimum(
                     p_min_pu[common_cols], nuclear_profiles[common_cols]
                 )
-        # Drop columns that are all zeros (default-value)
-        p_min_pu = p_min_pu.loc[:, (p_min_pu != 0.0).any()]
 
         p_max_pu = (
             profiles.pivot_table(values="p_max_pu", index="time", columns="bus")
@@ -1900,18 +2142,8 @@ def _add_conventional_thermal_capacities(
                 p_max_pu[common_cols] = np.minimum(
                     p_max_pu[common_cols], nuclear_profiles[common_cols]
                 )
-        # Drop columns that are all ones (default-value)
-        p_max_pu = p_max_pu.loc[:, (p_max_pu != 1.0).any()]
 
-        if not p_min_pu.empty:
-            index_name = n.links_t.p_min_pu.index.name
-            n.links_t.p_min_pu = pd.concat([n.links_t.p_min_pu, p_min_pu], axis=1)
-            n.links_t.p_min_pu.index.name = index_name
-
-        if not p_max_pu.empty:
-            index_name = n.links_t.p_max_pu.index.name
-            n.links_t.p_max_pu = pd.concat([n.links_t.p_max_pu, p_max_pu], axis=1)
-            n.links_t.p_max_pu.index.name = index_name
+        _add_pu_profiles(comp_t=n.links_t, p_min_pu=p_min_pu, p_max_pu=p_max_pu)
 
     # Remove non-expandable assets with no capacity
     remove_zero_capacity_non_extendable(
@@ -2331,7 +2563,7 @@ def _add_other_res_profiles(
     if carrier == "other-res-biomass":
         # adjust other-res-biomass profiles for MW_th as they are implemented as links
         profiles = profiles.assign(
-            p_set=lambda df: df.p_set.div(component_df.loc[asset_i].efficiency[0])
+            p_set=lambda df: df.p_set.div(component_df.loc[asset_i].efficiency.iloc[0])
         )
 
     p_set = (
@@ -9042,6 +9274,8 @@ if __name__ == "__main__":
         ext_carriers=snakemake.params.electricity.get("extendable_carriers", dict()),
         existing_capacities=existing_capacities,
         existing_efficiencies=existing_efficiencies,
+        pemmdb_capacities=pemmdb_capacities,
+        co2_price=co2_price,
     )
 
     if tyndp_hydro := [c for c in tyndp_renewable_carriers if c.startswith("hydro")]:
@@ -9057,6 +9291,7 @@ if __name__ == "__main__":
             costs=costs,
             pop_layout=pop_layout,
             spatial=spatial,
+            options=options,
         )
 
     add_h2_gas_infrastructure(